requirements.txt
flask
flask_cors
pymongo
pandas
plotly
python_dotenv
gunicorn
langchain
langchain_groq